+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Update sass-loader to get rid of warnings
 * Make all properties available as scss variables during the build
 * Add Template Element and Template Parts
+* Add studioLink and studioLinkEnabled
 
 
 ## [2.0.0] - 09.10.2025